更新3,完成国密算法库JNI接口编写及c语言源代码修改调试

+++ app/src/main/cpp/sm2_impl/sm2.c
@@ -1527,7 +1527,7 @@ int GM_SM2Encrypt(unsigned char * encData, unsigned long * ulEncDataLen, unsigne
 
         if(ulEncDataLen<512)
 	{
-		memset(encData,0,ulEncDataLen);
+		memset(encData, 0, (size_t) ulEncDataLen);
 	}
 	else
 	{
